Add files via upload
1. 
The results of the National Angler's Championship is stored in a text file. Each line of the 
file contains the identifier of the participant and the championship (strings without 
whitespace), and the list of the caught fish, which are stored as pairs: (the kind of the fish, 
the size of the fish). The kind of the fish is a string without whitespace, its size is a natural 
number. The data in a line are separated by whitespace. The lines of the text file are sorted 
according to the name of the championship. You can assume that the text file is correct.
An example for a line of the text file:
James BigLakeChampionship Tuna 50 Salmon 20 Sardine 5 Tuna 100
At which championship has the most fish been caught?